feat(float): Support min and max

+++ tests/float.test.js
@@ -1,11 +1,11 @@
 const tap = require('tap')
-const { int } = require('..')
+const { float } = require('..')
 
 tap.test('min and max', t => {
   let i = -1
 
   while (++i < 50) {
-    const result = int('foo', { min: 3, max: 5 })
+    const result = float('foo', { min: 3, max: 5 })
     t.assert(3 <= result && result <= 5)
   }
